requirements.txt
# Python 3.9

psycopg2-binary
python-dotenv
selenium
requests
pandas
numpy
